Refactor AI chat store and stabilize AI smoke tests
