+++ apps/ios/Flashcards/Flashcards/AI/AIChatStore+Dictation.swift
@@ -1,0 +1,143 @@
+import Foundation
+
+extension AIChatStore {
+    func toggleDictation() {
+        guard self.isChatInteractive else {
+            return
+        }
+        guard self.serverChatConfig.features.dictationEnabled || self.dictationState != .idle else {
+            return
+        }
+
+        switch self.dictationState {
+        case .idle:
+            guard self.hasExternalProviderConsent else {
+                self.showGeneralError(message: aiChatExternalProviderConsentRequiredMessage)
+                return
+            }
+            self.startDictation()
+        case .recording:
+            guard self.hasExternalProviderConsent else {
+                self.cancelDictation()
+                self.showGeneralError(message: aiChatExternalProviderConsentRequiredMessage)
+                return
+            }
+            self.finishDictation()
+        case .requestingPermission, .transcribing:
+            return
+        }
+    }
+
+    func cancelDictation() {
+        self.activeDictationTask?.cancel()
+        self.activeDictationTask = nil
+        self.voiceRecorder.cancelRecording()
+        self.dictationState = .idle
+        self.completedDictationTranscript = nil
+    }
+
+    func startDictation() {
+        if self.dictationState != .idle {
+            return
+        }
+
+        self.activeAlert = nil
+        self.completedDictationTranscript = nil
+        self.dictationState = .requestingPermission
+        self.activeDictationTask = Task { @MainActor in
+            defer {
+                self.activeDictationTask = nil
+            }
+
+            do {
+                try await self.voiceRecorder.startRecording()
+                self.dictationState = .recording
+            } catch is CancellationError {
+                self.dictationState = .idle
+            } catch let recorderError as AIChatVoiceRecorderError {
+                self.dictationState = .idle
+                self.handleStartDictationError(recorderError)
+            } catch {
+                self.dictationState = .idle
+                self.showGeneralError(error: error)
+            }
+        }
+    }
+
+    func finishDictation() {
+        if self.dictationState != .recording {
+            return
+        }
+
+        self.dictationState = .transcribing
+        self.activeDictationTask = Task { @MainActor in
+            defer {
+                self.activeDictationTask = nil
+            }
+
+            do {
+                guard self.hasExternalProviderConsent else {
+                    self.dictationState = .idle
+                    self.showGeneralError(message: aiChatExternalProviderConsentRequiredMessage)
+                    return
+                }
+                let session = try await self.flashcardsStore.cloudSessionForAI()
+                let recordedAudio = try await self.voiceRecorder.stopRecording()
+                defer {
+                    try? FileManager.default.removeItem(at: recordedAudio.fileUrl)
+                }
+
+                let transcription = try await self.audioTranscriber.transcribe(
+                    session: session,
+                    sessionId: self.chatSessionId.isEmpty ? nil : self.chatSessionId,
+                    recordedAudio: recordedAudio
+                )
+                self.chatSessionId = transcription.sessionId
+                self.schedulePersistCurrentState()
+                self.completedDictationTranscript = AIChatCompletedDictationTranscript(
+                    id: UUID().uuidString.lowercased(),
+                    transcript: transcription.text
+                )
+            } catch is CancellationError {
+            } catch let recorderError as AIChatVoiceRecorderError {
+                self.handleFinishDictationError(recorderError)
+            } catch let transcriptionError as AIChatTranscriptionError {
+                switch transcriptionError {
+                case .guestLimitReached:
+                    await self.appendStandaloneAssistantAccountUpgradePromptAndPersist(
+                        message: aiChatGuestQuotaReachedMessage,
+                        buttonTitle: aiChatGuestQuotaButtonTitle
+                    )
+                default:
+                    self.showGeneralError(error: transcriptionError)
+                }
+            } catch {
+                self.showGeneralError(error: error)
+            }
+
+            self.dictationState = .idle
+        }
+    }
+
+    func handleStartDictationError(_ error: AIChatVoiceRecorderError) {
+        switch error {
+        case .microphoneDenied:
+            return
+        case .microphoneBlocked:
+            self.showMicrophoneSettingsAlert()
+        default:
+            self.showGeneralError(error: error)
+        }
+    }
+
+    func handleFinishDictationError(_ error: AIChatVoiceRecorderError) {
+        switch error {
+        case .emptyRecording:
+            return
+        case .microphoneBlocked:
+            self.showMicrophoneSettingsAlert()
+        default:
+            self.showGeneralError(error: error)
+        }
+    }
+}
